automatically "fix" contract stage start times
When transitioning forward, if the stage that we are exiting's enter time does not match the previous stage's enter time, automatically adjust it to make them match. This can happen if, for example, the following steps are taken:

1. A user marks several steps in conductor, each, say 10 days apart. So, step one is marked complete January 1, step two is marked January 10, and step three is marked January 20.
2. A user reverts from step three to step two with the date box saying January 15.
3. The user marks step 2 complete complete on January 16.

When these steps occur in sequence, a gap of five days appears between step one and step two. This commit will automatically mark the step two enter date to be equal to the step one exit date.

The reason that we don't handle this on the revert side is because you can "revert to" a certain time right now, which leads to a common user behavior of reverting to the time you want to mark the step as complete and then immediately marking it as complete, or marking it as complete one day later. The path of least surprise in this case is to continue to allow the user to do this but to automatically back-adjust the reverted date instead of having him or her choose a date in the past and then find that conductor has reverted them to a different date altogether!